+++ stdlib/Random/src/DSFMT.jl
@@ -1,6 +1,6 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-module dSFMT
+module DSFMT
 
 import Base: copy, copy!, ==, hash
 using Base.GMP.MPZ